+++ auto_route/CHANGELOG.md
@@ -1,4 +1,7 @@
 # ChangeLog
+## [1.0.0-beta.10]
+- Fix generic parameters crash the generator
+- Add support for array type query params
 ## [1.0.0-beta.9]
 - Fix navigatorKey state is lost on hot reload
 - Add stackRouterOfIndex getter function to AutoTabsRouter